Refactor ServerHub apps view to use shared renderers
